Create Stack & Queue
Stack works on the principle of “Last-in, first-out”. Also, the inbuilt functions in Python make the code short and simple. To add an item to the top of the list, i.e., to push an item, we use append() function and to pop out an element we use pop() function.

Queue works on the principle of “First-in, first-out”.
enqueue - adds an element to the end of the queue.
dequeue - removes the element at the beginning of the queue.